+++ blocks/piwigo/index.asset.php
@@ -1,0 +1,5 @@
+<?php
+return [
+    'dependencies' => ['wp-blocks', 'wp-block-editor', 'wp-components', 'wp-compose', 'wp-element', 'wp-i18n', 'wp-server-side-render'],
+    'version' => '1.12.0',
+];